+++ src/any_agent/serving/server_a2a.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 import asyncio
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 
 import uvicorn
 from a2a.server.apps import A2AStarletteApplication
@@ -16,6 +16,7 @@ from any_agent.utils import run_async_in_sync
 from .agent_card import _get_agent_card
 from .agent_executor import AnyAgentExecutor
 from .envelope import prepare_agent_for_a2a, prepare_agent_for_a2a_async
+from .server_handle import ServerHandle
 
 if TYPE_CHECKING:
     from multiprocessing import Queue
@@ -77,7 +78,7 @@ async def serve_a2a_async(
     port: int,
     endpoint: str,
     log_level: str = "warning",
-) -> tuple[asyncio.Task[Any], uvicorn.Server]:
+) -> ServerHandle:
     """Provide an A2A server to be used in an event loop."""
     uv_server = _create_server(server, host, port, endpoint, log_level)
     task = asyncio.create_task(uv_server.serve())
@@ -86,7 +87,7 @@ async def serve_a2a_async(
     if port == 0:
         server_port = uv_server.servers[0].sockets[0].getsockname()[1]
         server.agent_card.url = f"http://{host}:{server_port}/{endpoint.lstrip('/')}"
-    return (task, uv_server)
+    return ServerHandle(task=task, server=uv_server)
 
 
 def serve_a2a(
@@ -103,11 +104,9 @@ def serve_a2a(
     # because the loop only keeps weak refs to tasks
     # https://docs.python.org/3/library/asyncio-task.html#asyncio.create_task
     async def run() -> None:
-        (task, uv_server) = await serve_a2a_async(
-            server, host, port, endpoint, log_level
-        )
+        server_handle = await serve_a2a_async(server, host, port, endpoint, log_level)
         if server_queue:
-            server_queue.put(uv_server.servers[0].sockets[0].getsockname()[1])
-        await task
+            server_queue.put(server_handle.port)
+        await server_handle.task
 
     return run_async_in_sync(run())
